Restructure PNP solver node and gather ground truth for pnp solver test
