+++ src/prm_vision/pnp_solver/test/spinning_in_place/ground_truth.csv
@@ -1,0 +1,135 @@
+986.70935,-62.69282,-204.93878
+978.38159,-57.83389,-203.93799
+969.77319,-56.90862,-202.37685
+969.77319,-56.90862,-202.37685
+969.77319,-56.90862,-202.37685
+970.62335,-57.20517,-202.56181
+969.86273,-57.39159,-202.39987
+970.69147,-57.18578,-202.58464
+970.69147,-57.18578,-202.58464
+969.81531,-57.41127,-202.38162
+969.77319,-56.90862,-202.37685
+969.77319,-56.90862,-202.37685
+969.77319,-56.90862,-202.37685
+979.45331,-57.89670,-204.67061
+969.86273,-57.39159,-202.39987
+970.62335,-57.20517,-202.56181
+968.75159,-57.10371,-202.15199
+969.81531,-57.41127,-202.38162
+970.62335,-57.20517,-202.56181
+970.69147,-57.18578,-202.58464
+969.86273,-57.39159,-202.39987
+997.27814,-107.08346,-215.67065
+901.60504,128.10960,-211.58504
+889.29828,56.91406,-202.59358
+1160.80505,-115.03479,-227.20671
+1171.77893,-63.32881,-248.51788
+942.03967,115.62262,-208.41472
+996.75183,173.37480,-218.02037
+901.02509,-9.53656,-206.69116
+899.30273,175.29085,-212.36519
+986.47247,252.70667,-226.12448
+947.79608,28.81864,-205.64848
+1004.44379,-21.19716,-210.51877
+1094.23486,-65.74242,-220.53679
+881.83160,100.69400,-206.20934
+1016.17053,-75.66067,-213.43684
+902.50635,101.41600,-202.17688
+949.43414,61.89783,-210.48598
+963.42444,231.63287,-225.26349
+934.63287,-53.01497,-210.84923
+926.31982,46.58491,-204.12262
+957.14209,4.39014,-206.95546
+908.39703,154.70502,-210.86407
+985.18140,119.53407,-230.66255
+1027.67529,224.11816,-223.14561
+962.39130,16.97910,-211.14290
+1057.95532,-25.11563,-226.50807
+853.94165,104.58428,-205.73206
+1018.57751,215.56398,-223.88283
+918.15717,92.39280,-205.49457
+917.64148,92.33153,-204.91850
+910.24921,91.50925,-203.48683
+925.64917,93.22914,-206.94774
+925.64917,93.22914,-206.94774
+918.15717,92.39280,-205.49457
+925.64917,93.22914,-206.94774
+917.64148,92.33153,-204.91850
+925.64917,93.22914,-206.94774
+917.64148,92.33153,-204.91850
+925.64917,93.22914,-206.94774
+910.45789,91.53773,-203.99548
+925.64917,93.22914,-206.94774
+917.64148,92.33153,-204.91850
+917.75092,92.34917,-205.40764
+918.15717,92.39280,-205.49457
+917.64148,92.33153,-204.91850
+917.64148,92.33153,-204.91850
+917.64148,92.33153,-204.91850
+917.64148,92.33153,-204.91850
+910.24921,91.50925,-203.48683
+918.15717,92.39280,-205.49457
+917.64148,92.33153,-204.91850
+917.64148,92.33153,-204.91850
+910.24921,91.50925,-203.48683
+918.15717,92.39280,-205.49457
+925.78448,93.03841,-206.99442
+917.64148,92.33153,-204.91850
+917.75092,92.34917,-205.40764
+910.24921,91.50925,-203.48683
+910.24921,91.50925,-203.48683
+918.15717,92.39280,-205.49457
+917.64148,92.33153,-204.91850
+910.24921,91.50925,-203.48683
+925.64917,93.22914,-206.94774
+910.24921,91.50925,-203.48683
+917.64148,92.33153,-204.91850
+917.64148,92.33153,-204.91850
+917.64148,92.33153,-204.91850
+918.15717,92.39280,-205.49457
+917.64148,92.33153,-204.91850
+917.64148,92.33153,-204.91850
+917.64148,92.33153,-204.91850
+925.64917,93.22914,-206.94774
+917.64148,92.33153,-204.91850
+917.64148,92.33153,-204.91850
+917.64148,92.33153,-204.91850
+917.64148,92.33153,-204.91850
+910.24921,91.50925,-203.48683
+917.64148,92.33153,-204.91850
+917.75092,92.34917,-205.40764
+910.24921,91.50925,-203.48683
+910.24921,91.50925,-203.48683
+917.64148,92.33153,-204.91850
+925.64917,93.22914,-206.94774
+917.64148,92.33153,-204.91850
+917.64148,92.33153,-204.91850
+917.64148,92.33153,-204.91850
+910.24921,91.50925,-203.48683
+917.64148,92.33153,-204.91850
+917.64148,92.33153,-204.91850
+917.64148,92.33153,-204.91850
+910.24921,91.50925,-203.48683
+917.64148,92.33153,-204.91850
+917.64148,92.33153,-204.91850
+917.64148,92.33153,-204.91850
+917.64148,92.33153,-204.91850
+918.15717,92.39280,-205.49457
+917.64148,92.33153,-204.91850
+910.24921,91.50925,-203.48683
+910.24921,91.50925,-203.48683
+917.75092,92.34917,-205.40764
+917.64148,92.33153,-204.91850
+917.64148,92.33153,-204.91850
+917.64148,92.33153,-204.91850
+917.64148,92.33153,-204.91850
+917.64148,92.33153,-204.91850
+917.64148,92.33153,-204.91850
+910.24921,91.50925,-203.48683
+925.64917,93.22914,-206.94774
+902.68530,90.67220,-202.01825
+909.88568,91.47081,-203.41048
+910.24921,91.50925,-203.48683
+917.64148,92.33153,-204.91850
+917.64148,92.33153,-204.91850
+933.67334,94.13274,-208.98383
