+++ src/prm_vision/pnp_solver/test/spinning_in_place/test_PNPSolver.cpp
@@ -1,0 +1,10 @@
+#include <gtest/gtest.h>
+#include <filesystem>
+#include "ament_index_cpp/get_package_share_directory.hpp"
+#include "PNPSolver.hpp"
+
+#include <fstream>
+#include <string>
+#include <vector>
+#include <sstream>
+
